requirements.txt
# AEON Platform Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI & ML Services
openai==1.3.7
replicate==0.22.0
anthropic==0.7.7

# Media Processing
Pillow==10.1.0
opencv-python==4.8.1.78
moviepy==1.0.3
ffmpeg-python==0.2.0

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
email-validator==2.1.0
phonenumbers==8.13.26

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.20.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Networking
websockets==12.0
sse-starlette==1.8.2

# Configuration
dynaconf==3.2.4

# Task Scheduling
apscheduler==3.10.4

# Text Processing
nltk==3.8.1
spacy==3.7.2

# API Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance
orjson==3.9.10
ujson==5.8.0

# Validation
jsonschema==4.20.0
cerberus==1.3.5

# Workflow Engine
networkx==3.2.1
graphviz==0.20.1

# Additional AI Services (optional)
elevenlabs==0.2.26
stability-sdk==0.8.4
